hbase applier better naming and cleenup
