+++ mysql-replicator-applier/src/test/java/com/booking/replication/applier/hbase/TableNameMapperTest.java
@@ -1,8 +1,5 @@
 package com.booking.replication.applier.hbase;
 
-import com.booking.replication.applier.hbase.schema.TableNameMapper;
-import org.junit.Test;
-
 import static org.junit.Assert.assertEquals;
 
 /**
